@@ -17,11 +17,11 @@
     },
     "devDependencies": {
         "@biomejs/biome": "1.9.4",
-        "@sveltejs/adapter-auto": "^3.2.4",
-        "@sveltejs/kit": "^2.0.0",
-        "@sveltejs/vite-plugin-svelte": "^3.0.0",
-        "svelte": "^4.2.7",
-        "vite": "^5.0.3",
+        "@sveltejs/adapter-auto": "^3.3.1",
+        "@sveltejs/kit": "^2.15.1",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0",
+        "svelte": "^5.16.5",
+        "vite": "^5.4.4",
         "vite-plugin-node-polyfills": "^0.22.0"
     }
 }
