@@ -1,6 +1,6 @@
 <script>
-export let fuel;
-import Furnace from "./Furnace.svelte";
+    let { fuel } = $props();
+    import Furnace from "./Furnace.svelte";
 </script>
 
 <Furnace input={null} {fuel} output={null} ticks={fuel.burnTime} />
