@@ -1,29 +1,61 @@
 <script>
-import { locale } from "$lib/stores";
-import { getTakeable, makeItemStack } from "$lib/utils";
-import CraftingRecipe from "../../CraftingRecipe.svelte";
-import Header from "../../Header.svelte";
-import ItemStackDetailDisplay from "../../ItemStackDetailDisplay.svelte";
+    import { locale } from "$lib/stores";
+    import { getTakeable, makeItemStack } from "$lib/utils";
+    import CraftingRecipe from "../../CraftingRecipe.svelte";
+    import Header from "../../Header.svelte";
+    import ItemStackDetailDisplay from "../../ItemStackDetailDisplay.svelte";
 
-import { page } from "$app/stores";
-import Body from "../../Body.svelte";
-import FuelRecipe from "../../FuelRecipe.svelte";
-import FurnaceRecipe from "../../FurnaceRecipe.svelte";
+    import { page } from "$app/stores";
+    import Body from "../../Body.svelte";
+    import FuelRecipe from "../../FuelRecipe.svelte";
+    import FurnaceRecipe from "../../FurnaceRecipe.svelte";
 
-import { db } from "$lib/db";
-import { liveQuery } from "dexie";
+    import { db } from "$lib/db";
+    import { liveQuery } from "dexie";
 
-$: itemStack = liveQuery(async () => {
-    return await makeItemStack(await getTakeable($page.params.item));
-});
+    let itemStack = $derived(
+        liveQuery(
+            async () => {
+                return await makeItemStack(
+                    await getTakeable($page.params.item),
+                );
+            },
+            {},
+            $page.params.item,
+        ),
+    );
 
-$: craftingRecipes = liveQuery(() => db.craftingRecipes.where("usedItemsFullIds").equals($page.params.item).toArray());
+    let craftingRecipes = $derived(
+        liveQuery(
+            () =>
+                db.craftingRecipes
+                    .where("usedItemsFullIds")
+                    .equals($page.params.item)
+                    .toArray(),
+            {},
+            $page.params.item,
+        ),
+    );
 
-$: furnaceRecipes = liveQuery(() => db.furnaceRecipes.where("usedItem.fullId").equals($page.params.item).toArray(), { initialValue: [] });
+    let furnaceRecipes = $derived(
+        liveQuery(
+            () =>
+                db.furnaceRecipes
+                    .where("usedItem.fullId")
+                    .equals($page.params.item)
+                    .toArray(),
+            { initialValue: [] },
+            $page.params.item,
+        ),
+    );
 
-const name = liveQuery(async () => {
-    return await $itemStack?.getName($locale);
-});
+    let name = liveQuery(
+        async () => {
+            return await $itemStack?.getName($locale);
+        },
+        {},
+        [$itemStack, $locale],
+    );
 </script>
 
 <svelte:head>
