@@ -1,19 +1,19 @@
 <script>
-import Body from "../Body.svelte";
-import Header from "../Header.svelte";
-import SearchableItemList from "../SearchableItemList.svelte";
+    import Body from "../Body.svelte";
+    import Header from "../Header.svelte";
+    import SearchableItemList from "../SearchableItemList.svelte";
 
-import { db } from "$lib/db";
-import { mergeByKey } from "$lib/utils";
-import { liveQuery } from "dexie";
+    import { db } from "$lib/db";
+    import { mergeByKey } from "$lib/utils";
+    import { liveQuery } from "dexie";
 
-const fuels = liveQuery(async () => {
-    const items = (await db.items.toArray()) || [];
-    const blockStates = (await db.blockstates.toArray()) || [];
-    const takeables = items.concat(blockStates);
-    const fuels = takeables.filter((takeable) => takeable.isFuel);
-    return fuels;
-});
+    let fuels = liveQuery(async () => {
+        const items = (await db.items.toArray()) || [];
+        const blockStates = (await db.blockstates.toArray()) || [];
+        const takeables = items.concat(blockStates);
+        const fuels = takeables.filter((takeable) => takeable.isFuel);
+        return fuels;
+    });
 </script>
 
 <svelte:head>
@@ -28,11 +28,13 @@ const fuels = liveQuery(async () => {
     <h2>Fuels</h2>
 
     <div class="center">
-        <SearchableItemList takeables={mergeByKey(
-            $fuels || [],
-            (obj) => obj.fullId,
-            ([a, b]) => b.source === "jar",
-        )} />
+        <SearchableItemList
+            takeables={mergeByKey(
+                $fuels || [],
+                (obj) => obj.fullId,
+                ([a, b]) => b.source === "jar",
+            )}
+        />
         {#if !$fuels?.length}
             <p>There are no fuels.</p>
         {/if}
